Remove dependency on vim-rhubarb
Because we've rewritten all the GitHub API code, we were only using
rhubarb for the homepage_from_url function.
We've replicated this function, and the fugitive gbrowse_handler.

This plugin can now be used on its own (with fugitive) and now directly
supports :GBrowse for GitHub

If vim-rhubarb is installed, either this repo or rhubarb will handle
GBrowse, whichever comes first.